11/7 mod, strtrim and split otw

+++ ft_split.c
@@ -3,26 +3,26 @@
 /*                                                        :::      ::::::::   */
 /*   ft_split.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jihyjeon <jihyjeon@student.42seoul.>       +#+  +:+       +#+        */
+/*   By: jihyjeon < jihyjeon@student.42seoul.kr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 18:43:18 by jihyjeon          #+#    #+#             */
-/*   Updated: 2023/10/19 18:44:07 by jihyjeon         ###   ########.fr       */
+/*   Updated: 2023/11/07 22:28:45 by jihyjeon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ch_sep(char sc, char c)
+size_t	ch_sep(char sc, char c)
 {
 	if (c == sc)
 		return (1);
 	return (0);
 }
 
-int	s_cnt(const char *str, char chr)
+size_t	s_cnt(const char *str, char chr)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 1;
@@ -35,9 +35,9 @@ int	s_cnt(const char *str, char chr)
 	return (j);
 }
 
-int	s_len(const char *str, char c)
+size_t	s_len(const char *str, char c)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
@@ -51,8 +51,8 @@ int	s_len(const char *str, char c)
 
 char	*c_arr(const char *str, char c)
 {
-	int		i;
-	int		len;
+	size_t	i;
+	size_t	len;
 	char	*new;
 
 	len = s_len(str, c);
@@ -72,8 +72,8 @@ char	*c_arr(const char *str, char c)
 char	**ft_split(char const *s, char c)
 {
 	char	**arr;
-	int		i;
-	int		nos;
+	size_t	i;
+	size_t	nos;
 
 	nos = s_cnt(s, c);
 	arr = (char **)malloc(sizeof(char *) * (nos + 1));
